move handlers into internal
